KeyWordEditText SingLine Set / RecyclerView onBinding Data parsing

+++ app/src/main/java/com/example/abled_food_connect/data/MainFragmentItemData.kt
@@ -28,7 +28,7 @@ data class MainFragmentItemData(
     @SerializedName("hostName")
     var hostName:String,
     @SerializedName("roomStatus")
-    var roomStatus:Int
+    var roomStatus:Double
 
 ){
 private fun CreateRoom(){
